+++ src_code/image2x-Extension-QT/mainwindow.h
@@ -83,7 +83,7 @@ public:
     MainWindow(QWidget *parent = nullptr);
     void changeEvent(QEvent *e);
     //=======================
-    QString VERSION = "v0.3.2";
+    QString VERSION = "v0.3.5";
     bool isBetaVer = false;
     QString LastStableVer = "v3.31.13";
     QString LastBetaVer = "v3.41.02-beta";